refactor(tests): Move slow moving tests to tests/nocover
